config changes and reset password changes

+++ app/controllers/attendances_controller.rb
@@ -33,6 +33,9 @@ class AttendancesController < ApplicationController
   # GET /attendances/1/edit
   def edit
     @attendance = Attendance.find(params[:id])
+    if(@attendance.approval_status)
+      redirect_to "/faultyaccess.html"
+    end
   end
 
   # POST /attendances
@@ -41,6 +44,9 @@ class AttendancesController < ApplicationController
     @attendance = Attendance.new(params[:attendance])
     @attendance.user = current_user
     @attendance.end_date= @attendance.start_date + @attendance.days.days
+    if(@attendance.approval_status)
+      redirect_to "/faultyaccess.html"
+    end
 
     respond_to do |format|
       if @attendance.save
@@ -58,6 +64,9 @@ class AttendancesController < ApplicationController
   def update
     @attendance = Attendance.find(params[:id])
     @attendance.end_date= @attendance.start_date + @attendance.days.days
+    if(@attendance.approval_status)
+      redirect_to "/faultyaccess.html"
+    end
 
     respond_to do |format|
       if @attendance.update_attributes(params[:attendance])
@@ -74,6 +83,9 @@ class AttendancesController < ApplicationController
   # DELETE /attendances/1.json
   def destroy
     @attendance = Attendance.find(params[:id])
+    if(@attendance.approval_status)
+      redirect_to "/faultyaccess.html"
+    end
     if @attendance.destroy
       redirect_to :controller => 'dashboard', :notice => "Attendance Deleted successfully"
     end
